Updated Pipeline Designer to be Production Cookbook.
I created new menus for maya that are much simpler - one menu
with all the functions.  I created tests for build, publish, and render
to make sure something exists, and if not throw a popup with instructions.

I also added a copy test code button to pipeline designer that will
allow you to copy code into your software of choice so you can test it
live while working on a button or preflight.